requirements.txt

msgpack-python
ez_setup # Fixes pyzmq ez-setup dependency
gevent
pyzmq
leveldb
